Crosswalks and demo scene updated
The demo scene has been changed to represent a 4 way intersection.

A basic crosswalk has been built, and a timer and switching mechanism
has been put in place. The timer function built into the crosswalks
should be replaced by logic in the traffic lights later on.

I also added a stress test with a spawner that creates 1000 pedestrians
and gives them all the same path. This may not be a great test right now
as they're all using similar paths, but it was fun to make.

Some issues remain with the crosswalks, such as people getting stuck at
waiting locations at crosswalks.

I'd like to make the crosswalks auto-generate from the Game Object the
script is placed on, and that will probably be my focus next week.